Update README.md
An outbreak of catastrophic proportions
known as the COVID-19 has had a pronounced
effect on people’s livelihood. In the midst
of this, social media platforms are noticeably
some of the prominent means to spread awareness,
keep people connected and entertained.
In this paper, topic modelling is implemented
to distribute the data set containing posts extracted
from Twitter into five topics, and sentimental
analysis and emotion detection analysis
is conducted. This provides insights on the
general perspectives and emotions of users in
this circumstance. The results indicate an overall
positive outlook with only less than 36%
representing an average negative sentiment.